app/vmselect: support clusternative select API for single

+++ app/vmselect/clusternative/vmselect.go
@@ -8,7 +8,6 @@ import (
 	"github.com/VictoriaMetrics/VictoriaMetrics/app/vmselect/netstorage"
 	"github.com/VictoriaMetrics/VictoriaMetrics/app/vmselect/searchutil"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/cgroup"
-	"github.com/VictoriaMetrics/VictoriaMetrics/lib/httputil"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/querytracer"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/storage"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/vmselectapi"
@@ -48,9 +47,11 @@ func NewVMSelectServer(addr string) (*vmselectapi.Server, error) {
 type vmstorageAPI struct{}
 
 func (api *vmstorageAPI) InitSearch(qt *querytracer.Tracer, sq *storage.SearchQuery, deadline uint64) (vmselectapi.BlockIterator, error) {
-	denyPartialResponse := httputil.GetDenyPartialResponse(nil)
+	if !netstorage.IsSameTenant(sq.AccountID, sq.ProjectID) {
+		return nil, nil
+	}
 	dl := searchutil.DeadlineFromTimestamp(deadline)
-	bi := newBlockIterator(qt, denyPartialResponse, sq, dl)
+	bi := newBlockIterator(qt, sq, dl)
 	return bi, nil
 }
 
@@ -60,66 +61,89 @@ func (api *vmstorageAPI) Tenants(qt *querytracer.Tracer, tr storage.TimeRange, d
 }
 
 func (api *vmstorageAPI) SearchMetricNames(qt *querytracer.Tracer, sq *storage.SearchQuery, deadline uint64) ([]string, error) {
-	denyPartialResponse := httputil.GetDenyPartialResponse(nil)
+	if !netstorage.IsSameTenant(sq.AccountID, sq.ProjectID) {
+		return nil, nil
+	}
 	dl := searchutil.DeadlineFromTimestamp(deadline)
-	metricNames, _, err := netstorage.SearchMetricNames(qt, denyPartialResponse, sq, dl)
+	metricNames, err := netstorage.SearchMetricNames(qt, sq, dl)
 	return metricNames, err
 }
 
 func (api *vmstorageAPI) LabelValues(qt *querytracer.Tracer, sq *storage.SearchQuery, labelName string, maxLabelValues int, deadline uint64) ([]string, error) {
-	denyPartialResponse := httputil.GetDenyPartialResponse(nil)
+	if !netstorage.IsSameTenant(sq.AccountID, sq.ProjectID) {
+		return nil, nil
+	}
 	dl := searchutil.DeadlineFromTimestamp(deadline)
-	labelValues, _, err := netstorage.LabelValues(qt, denyPartialResponse, labelName, sq, maxLabelValues, dl)
+	labelValues, err := netstorage.LabelValues(qt, labelName, sq, maxLabelValues, dl)
 	return labelValues, err
 }
 
 func (api *vmstorageAPI) TagValueSuffixes(qt *querytracer.Tracer, accountID, projectID uint32, tr storage.TimeRange, tagKey, tagValuePrefix string, delimiter byte,
 	maxSuffixes int, deadline uint64) ([]string, error) {
-	denyPartialResponse := httputil.GetDenyPartialResponse(nil)
+	if !netstorage.IsSameTenant(accountID, projectID) {
+		return nil, nil
+	}
 	dl := searchutil.DeadlineFromTimestamp(deadline)
-	suffixes, _, err := netstorage.TagValueSuffixes(qt, accountID, projectID, denyPartialResponse, tr, tagKey, tagValuePrefix, delimiter, maxSuffixes, dl)
+	suffixes, err := netstorage.TagValueSuffixes(qt, tr, tagKey, tagValuePrefix, delimiter, maxSuffixes, dl)
 	return suffixes, err
 }
 
 func (api *vmstorageAPI) LabelNames(qt *querytracer.Tracer, sq *storage.SearchQuery, maxLabelNames int, deadline uint64) ([]string, error) {
-	denyPartialResponse := httputil.GetDenyPartialResponse(nil)
+	if !netstorage.IsSameTenant(sq.AccountID, sq.ProjectID) {
+		return nil, nil
+	}
 	dl := searchutil.DeadlineFromTimestamp(deadline)
-	labelNames, _, err := netstorage.LabelNames(qt, denyPartialResponse, sq, maxLabelNames, dl)
+	labelNames, err := netstorage.LabelNames(qt, sq, maxLabelNames, dl)
 	return labelNames, err
 }
 
 func (api *vmstorageAPI) SeriesCount(qt *querytracer.Tracer, accountID, projectID uint32, deadline uint64) (uint64, error) {
-	denyPartialResponse := httputil.GetDenyPartialResponse(nil)
+	if !netstorage.IsSameTenant(accountID, projectID) {
+		return 0, nil
+	}
 	dl := searchutil.DeadlineFromTimestamp(deadline)
-	seriesCount, _, err := netstorage.SeriesCount(qt, accountID, projectID, denyPartialResponse, dl)
+	seriesCount, err := netstorage.SeriesCount(qt, dl)
 	return seriesCount, err
 }
 
 func (api *vmstorageAPI) TSDBStatus(qt *querytracer.Tracer, sq *storage.SearchQuery, focusLabel string, topN int, deadline uint64) (*storage.TSDBStatus, error) {
-	denyPartialResponse := httputil.GetDenyPartialResponse(nil)
+	if !netstorage.IsSameTenant(sq.AccountID, sq.ProjectID) {
+		return nil, nil
+	}
 	dl := searchutil.DeadlineFromTimestamp(deadline)
-	tsdbStatus, _, err := netstorage.TSDBStatus(qt, denyPartialResponse, sq, focusLabel, topN, dl)
+	tsdbStatus, err := netstorage.TSDBStatus(qt, sq, focusLabel, topN, dl)
 	return tsdbStatus, err
 }
 
 func (api *vmstorageAPI) DeleteSeries(qt *querytracer.Tracer, sq *storage.SearchQuery, deadline uint64) (int, error) {
+	if !netstorage.IsSameTenant(sq.AccountID, sq.ProjectID) {
+		return 0, nil
+	}
 	dl := searchutil.DeadlineFromTimestamp(deadline)
 	return netstorage.DeleteSeries(qt, sq, dl)
 }
 
 func (api *vmstorageAPI) RegisterMetricNames(qt *querytracer.Tracer, mrs []storage.MetricRow, deadline uint64) error {
+	dest := mrs[:0]
+	for _, m := range mrs {
+		if netstorage.HasSameTenantPrefix(m.MetricNameRaw) {
+			dest = append(dest, m)
+		}
+	}
+	mrs = dest
 	dl := searchutil.DeadlineFromTimestamp(deadline)
 	return netstorage.RegisterMetricNames(qt, mrs, dl)
 }
 
 func (api *vmstorageAPI) ResetMetricNamesUsageStats(qt *querytracer.Tracer, deadline uint64) error {
-	dl := searchutil.DeadlineFromTimestamp(deadline)
-	return netstorage.ResetMetricNamesStats(qt, dl)
+	return netstorage.ResetMetricNamesStats(qt)
 }
 
 func (api *vmstorageAPI) GetMetricNamesUsageStats(qt *querytracer.Tracer, tt *storage.TenantToken, le, limit int, matchPattern string, deadline uint64) (storage.MetricNamesStatsResponse, error) {
-	dl := searchutil.DeadlineFromTimestamp(deadline)
-	return netstorage.GetMetricNamesStats(qt, tt, le, limit, matchPattern, dl)
+	if !netstorage.IsSameTenant(tt.AccountID, tt.ProjectID) {
+		return storage.MetricNamesStatsResponse{}, nil
+	}
+	return netstorage.GetMetricNamesStats(qt, le, limit, matchPattern)
 }
 
 // blockIterator implements vmselectapi.BlockIterator
@@ -134,12 +158,12 @@ type workItem struct {
 	doneCh chan struct{}
 }
 
-func newBlockIterator(qt *querytracer.Tracer, denyPartialResponse bool, sq *storage.SearchQuery, deadline searchutil.Deadline) *blockIterator {
+func newBlockIterator(qt *querytracer.Tracer, sq *storage.SearchQuery, deadline searchutil.Deadline) *blockIterator {
 	var bi blockIterator
 	bi.workCh = make(chan workItem, 16)
 	bi.wg.Add(1)
 	go func() {
-		_, err := netstorage.ProcessBlocks(qt, denyPartialResponse, sq, func(mb *storage.MetricBlock, _ uint) error {
+		err := netstorage.ProcessBlocks(qt, sq, func(mb *storage.MetricBlock) error {
 			wi := workItem{
 				mb:     mb,
 				doneCh: make(chan struct{}),
